requirements.txt
# Core ML
torch>=2.0.0
pytorch-lightning>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0

# Chemistry
rdkit>=2023.3.1

# Experiment Tracking
comet-ml>=3.35.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web App
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0